requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1
python-dotenv==1.0.0
redis==4.6.0
flask-caching==2.1.0
